pyproject.toml
[project]
name = "foreign-languages-cards-api"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13.7"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "fastapi>=0.118.0",
    "pydantic-settings>=2.11.0",
    "sqlalchemy[asyncio]>=2.0.40",
    "uvicorn[standard]>=0.34.2",
]
